+++ CHANGELOG.md
@@ -273,3 +273,4 @@
 - 2024-01-30: Impl descend, bettercontact, lead, LICENSE, +3 query
 - 2024-01-31: Refactor bettercontact, README, daemon, lib, +5
 - 2024-02-01: Impl bettercontact, discovery, LICENSE, lib, +3 query
+- 2024-02-01: Wire bettercontact, users, campaign, lib, +3